Move self destruct command definition code to module
